fix(code): add useMediaQuery hook
